Seperate logic from the dom and split the code
